requirements.txt
numpy==1.14.5
pandas==0.23.0
tqdm==4.23.4
scipy==1.1.0
torch==0.4.0
pyproj
nupic.bindings
